UGENE-7769 UGENECL has an execution error

+++ src/plugins_3rdparty/primer3/src/Primer3Plugin.cpp
@@ -77,7 +77,9 @@ Primer3Plugin::Primer3Plugin()
 }
 
 Primer3Plugin::~Primer3Plugin() {
-    viewCtx->deleteLater();
+    if (viewCtx != nullptr) {
+        viewCtx->deleteLater();
+    }
 }
 
 Primer3ADVContext::Primer3ADVContext(QObject* p)
